feat:use cmd+k to search
