+++ package.json
@@ -55,6 +55,7 @@
     "history": "5.3.0",
     "i18next": "20.6.1",
     "jsbi": "3.2.5",
+    "kbar": "^0.1.0-beta.45",
     "lint-staged": "^13.2.3",
     "lodash.debounce": "4.0.8",
     "lucide-react": "^0.525.0",
