+++ src/components/Search/utils.ts
@@ -1,5 +1,11 @@
-import { SearchResultType, AggregateSearchResult } from '../../services/ExplorerService'
-import { TYPE_ID_CODE_HASH } from '../../constants/common'
+import { explorerService, Response, SearchResultType, type AggregateSearchResult } from '../../services/ExplorerService'
+import { getReverseAddresses } from '../../services/DidService'
+import { ethToCKb as DidEthToCkb } from '../../utils/did'
+import { addPrefixForHash, containSpecialChar } from '../../utils/string'
+import { HttpErrorCode, SearchFailType, TYPE_ID_CODE_HASH } from '../../constants/common'
+import { isChainTypeError } from '../../utils/chain'
+import { isRequestError } from '../../utils/error'
+import { isValidBTCAddress } from '../../utils/bitcoin'
 
 export const getURLByAggregateSearchResult = (result: AggregateSearchResult) => {
   const { type, attributes } = result
@@ -105,5 +111,94 @@ export const getDisplayNameByAggregateSearchResult = (result: AggregateSearchRes
   }
   if (type === SearchResultType.FiberGraphNode) {
     return attributes.peerId
+  }
+}
+
+export const ALLOW_SEARCH_TYPES = [
+  SearchResultType.Address,
+  SearchResultType.Block,
+  SearchResultType.BtcTx,
+  SearchResultType.LockHash,
+  SearchResultType.LockScript,
+  SearchResultType.Transaction,
+  SearchResultType.TypeScript,
+  SearchResultType.TokenCollection,
+  SearchResultType.TokenItem,
+  SearchResultType.UDT,
+  SearchResultType.DID,
+  SearchResultType.BtcAddress,
+  SearchResultType.FiberGraphNode,
+]
+
+export async function fetchAggregateSearchResult(searchValue: string): Promise<AggregateSearchResult[]> {
+  let results = await explorerService.api
+    .fetchAggregateSearchResult(addPrefixForHash(searchValue))
+    .then(res => res.data)
+    .catch(() => [] as AggregateSearchResult[])
+
+  if (/\w*\.bit$/.test(searchValue)) {
+    // search .bit name
+    const list = await getReverseAddresses(searchValue)
+    const ETH_COIN_TYPE = '60'
+    const ethAddr = list?.find(item => item.key_info.coin_type === ETH_COIN_TYPE)
+    if (ethAddr) {
+      const ckbAddr = DidEthToCkb(ethAddr.key_info.key)
+      results = [
+        ...results,
+        {
+          id: Math.random(),
+          type: SearchResultType.DID,
+          attributes: {
+            did: searchValue,
+            address: ckbAddr,
+          },
+        },
+      ]
+    }
+  }
+
+  return results
+}
+
+export const getURLBySearchValue = async (searchValue: string) => {
+  // check whether is btc address
+  if (isValidBTCAddress(searchValue)) {
+    return `/address/${searchValue}`
+  }
+  if (/\w*\.bit$/.test(searchValue)) {
+    // search .bit name
+    const list = await getReverseAddresses(searchValue)
+    const ETH_COIN_TYPE = '60'
+    const ethAddr = list?.find(item => item.key_info.coin_type === ETH_COIN_TYPE)
+    if (ethAddr) {
+      const ckbAddr = DidEthToCkb(ethAddr.key_info.key)
+      return `/address/${ckbAddr}`
+    }
+  }
+  // TODO: Is this replace needed?
+  const query = addPrefixForHash(searchValue).replace(',', '')
+  if (!query || containSpecialChar(query)) {
+    return `/search/fail?q=${query}`
+  }
+  if (isChainTypeError(query)) {
+    return `/search/fail?type=${SearchFailType.CHAIN_ERROR}&q=${query}`
+  }
+
+  try {
+    const data = await fetchAggregateSearchResult(addPrefixForHash(query))
+    return getURLByAggregateSearchResult(data[0])
+  } catch (error) {
+    if (
+      isRequestError(error) &&
+      error.response?.data &&
+      error.response.status === 404 &&
+      (error.response.data as Response.Error[]).find(
+        (errorData: Response.Error) => errorData.code === HttpErrorCode.NOT_FOUND_ADDRESS,
+      )
+    ) {
+      return `/address/${query}`
+    }
+
+    return `/search/fail?q=${query}`
   }
 }
